refactor route-browser
move routes-fetching logic to view from table component

remove accidentally created file
